app/requirements.txt
Flask
mysqlclient
flask-mail
flask-weasyprint
passlib
pycryptodome
Flask-WTF
email-validator
flask-login
passlib[bcrypt]